test(ui): mock nativewind + react-native-svg in RN tests; remove RN->web mock
